requirements.txt
colorama
discord
requests
bs4
mysql-connector
tpblite
psutil